refactor: HeroUnits section
